// src/rolling-text/types.ts
// styles props

export interface NumberProps {
  fontSize: number;
  backgroundColor?: string;
  color?: string;
}

export interface TextProps {
  fontSize: number;
}

export interface FocusedProps {
  fontSize: number;
  backgroundColor?: string;
}

export interface MainContainerProps {
  width: string;
  height: string;
  fontSize: number;
  backgroundColor?: string;
}

export interface BlurProps {
  backgroundColor?: string;
}

export interface UnitBoxProps {
  backgroundColor?: string;
  color?: string;
}